# record-server-wsdl/src/main/swagger/swagger-record-server.yml
swagger: '2.0'
info:
  version: 2.0.3
  title: Recordserver
  description: TODO
host: recordserver.trafalgar.ws:8080
basePath: /v2
schemes:
- http
produces:
- application/json
paths:
  /ship/{ship}/maps/{map}:
    get:
      description: TODO
      operationId: getByShipAndMap
      parameters:
      - name: map
        in: path
        required: true
        type: string
        description: TODO
      - name: ship
        in: path
        required: true
        type: string
        description: TODO
      responses:
        200:
          description: TODO
          schema:
            $ref: '#/definitions/RankingEntry'
  /id/{id}:
    get:
      description: TODO
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        type: string
        description: TODO
      responses:
        200:
          description: TODO
          schema:
            $ref: '#/definitions/MapEntry'
  /ranking:
    post:
      description: TODO
      operationId: postRanking
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: New ranking request.
        required: true
        schema:
          $ref: "#/definitions/NewMapEntryRequest"
      responses:
        200:
          description: TODO
          schema:
            $ref: '#/definitions/NewMapEntryResponse'
definitions:
  RankingEntry:
    type: object
    properties:
      rankingEntry:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            position:
              type: integer
              format: int32
            time:
              type: number
              format: float
            username:
              type: string
  NewMapEntryRequest:
    type: object
    properties:
      payload:
        type: string
  NewMapEntryResponse:
    type: object
    properties:
      id:
        type: string
      position:
        type: integer
        format: int32
      status:
        type: string
      verified:
        type: boolean
      username:
        type: string
  MapEntry:
    type: object
    properties:
      id:
        type: string
      payload:
        type: string